Resolve name conflict in event handlers with "xx".
this.xx was the scope of an element in an event handler. This
conflicts with the global "xx" entry point. Event handler seams
to be called with with(el) { eventhandler(event) }.
Now each html element with an attached scope can access the scope
with "el.$scope" which is in a handler: "this.$scope" or short
"$scope" (undocumented feature? Cant find documentation.)